refactor(collector-core): split event_bus.rs into event_bus/ submodules
Pure mechanical module extraction with no logic, behavior, or public-API
changes. The monolithic event_bus.rs (2630 lines) is split into:

- event_bus/mod.rs (65 lines): module doc, EventBus trait, glob re-exports
  preserving the full public API surface
- event_bus/types.rs (462 lines): config, subscription, filter, correlation,
  bus event, and statistics types plus CorrelationMetadata impl
- event_bus/local_bus.rs (2137 lines): LocalEventBus (single cohesive impl
  of one type) plus co-located tests that exercise its private associated
  functions (topic_matches_pattern, validate_single_pattern)

local_bus.rs remains large because it is one type's single impl (~975 lines
non-test) plus tests that must stay co-located to reach private items via
super::*; sharding one impl or moving the tests would change visibility.

Test count unchanged (147 lib tests, 12 in event_bus).
